feat: create new pages
