+++ README.md
@@ -1,0 +1,5 @@
+criar logica nas paginas de projeto para botao de fechar quando descer a pagina se torne fixed ou static
+
+criar botao de voltar ao topo
+
+criar os perfils de cores de acessibilidade
