+++ index.html
@@ -233,12 +233,7 @@
                 </details>
             </div>
         </section>
-        <section class="container tutorials">
-            <h1>Olá, o que vamos aprender hoje?</h1>
-            <p>Com o intuito de compartilhar conhecimento, criei esse espaço onde você pode aproveitar e aprender sobre os mais diversos temas da tecnólogia que vão desde o Design ate a programação.</p>
-            <div class="grid-cards"></div>
-        </section>
-        
+
     </main>
 
     <footer>
