+++ public/assets/projects/pop-cinema/pop-cinema.json
@@ -1,0 +1,48 @@
+{
+    "project": [
+        {
+            "title": "breve descricao",
+            "client": "para que foi feito",
+            "link-figma": "",
+            "link-github": "",
+            "link-aplication": "",
+            "img-cover": "",
+
+            "description-project": "",
+            "goal-project": "",
+
+            "link-iframe-figma-mobile": "",
+            "link-iframe-figma-desktop": "",
+
+            "study-case-process": "como fez para quem porque",
+            "img-case-1": "",
+            "img-case-2": "",
+            "img-case-3": "",
+            "img-case-4": "",
+            "img-case-5": "",
+            "img-case-6": "",
+            "img-case-7": "",
+            "img-case-8": "",
+            "img-case-9": "",
+            "img-case-10": "",
+            "img-case-11": "",
+            "img-case-12": "",
+
+            "style-guide-process": "",
+            "img-style-guide-1": "",
+            "img-style-guide-2": "",
+            "img-style-guide-3": "",
+            "img-style-guide-4": "",
+            "img-style-guide-5": "",
+
+            "mockups-process": "descricao telas depois prontas",
+            "img-mockup-1": "",
+            "img-mockup-2": "",
+            "img-mockup-3": "",
+            "img-mockup-4": "",
+
+            "chanllenge-process": ""
+
+        }
+]
+}
