+++ public/assets/projects/template-study-case.json
