+++ src/pages/pop-cinema.html
@@ -1,0 +1,101 @@
+<!DOCTYPE html>
+<html lang="pt_BR" ui-theme="light">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta name="description"
+        content="Dener Garcia é um UX Engineer que cria interfaces e aplicações incriveis, conheça meu portfolio de UX/UI Designer, Desenvolvedor Front-end e Mobile">
+    <meta name="robots" content="index">
+
+    <meta property=”og:title” content=”Portifólio de Dener Garcia | UX Engineer“>
+    <meta property="og:type" content="website">
+    <meta property="og:desctiption"
+        content="Profissional que atua na interseção entre design de interface de usuário (UI) e desenvolvimento de front-end, focando em criar experiências digitais eficazes e agradáveis para os usuários.">
+    <meta property="og:locale" content="pt_BR" />
+    <meta property="og:url" content="https://dener.criarbr.com/" />
+    <meta property="og:image" content="https://dener.criarbr.com/public/assets/dener-garcia.opti.jpg" />
+    <meta property="og:image:width" content="600" />
+    <meta property="og:image:height" content="800" />
+
+    <link rel="canonical" href="https://dener.criarbr.com/">
+    <link rel="sitemap" href="sitemap.xml">
+    <link rel="icon" type="image/svg+xml" href="../public/assets/avatar.ico">
+
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link
+        href="https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,300;0,400;0,500;0,700;1,300&display=swap&family=Audiowide&display=swap"
+        rel="stylesheet" />
+    <link rel="stylesheet" type="text/css" href="/src/styles/style.css" />
+
+    <script src="/src/functions/youtube-call/figma-tutorials.js" type="module" defer></script>
+    <script src="https://unpkg.com/@phosphor-icons/web" defer></script>
+
+    <title>Dener Garcia | Blog</title>
+</head>
+
+<body>
+
+
+
+    <header>
+        <a href="/index.html" class="logo">&lt; DenerGarcia /&gt;</a>
+        <div class="settings-ui">
+            <i class="ph ph-person-arms-spread"></i>
+            <button>Acessibilidade</button>
+            <div class="ui-options d-none">
+
+                <div class="option">
+                    <label for="light_theme"><i class="ph ph-sun-dim"> </i>Tema Claro</label>
+                    <input type="radio" id="light_theme" name="ui-theme" value="light-theme" checked>
+                </div>
+                <div class="option">
+                    <label for="dark_theme"> <i class="ph ph-moon-stars"> </i>Tema Escuro</label>
+                    <input type="radio" id="dark_theme" name="ui-theme" value="dark-theme">
+                </div>
+                <div class="option">
+                    <label for="high_contrast"><i class="ph-fill ph-circle-half-tilt"> </i>Alto Contraste</label>
+                    <input type="radio" id="high_contrast" name="ui-theme" value="high-contrast">
+                </div>
+            </div>
+        </div>
+        <a href="/index.html#projects" class="btn-return" aria-label="Menu do site">
+            <span></span>
+            <span></span>
+        </a>
+    </header>
+
+    <main>
+        <section class="container tutorials">
+            <h1>Olá, o que vamos aprender hoje?</h1>
+            <p>Com o intuito de compartilhar conhecimento, criei esse espaço onde você pode aproveitar e aprender sobre os mais diversos temas da tecnólogia que vão desde o Design ate a programação.</p>
+            h1>Olá, o que vamos aprender hoje?</h1>
+            <p>Com o intuito de compartilhar conhecimento, criei esse espaço onde você pode aproveitar e aprender sobre os mais diversos temas da tecnólogia que vão desde o Design ate a programação.</p>h1>Olá, o que vamos aprender hoje?</h1>
+            <p>Com o intuito de compartilhar conhecimento, criei esse espaço onde você pode aproveitar e aprender sobre os mais diversos temas da tecnólogia que vão desde o Design ate a programação.</p>h1>Olá, o que vamos aprender hoje?</h1>
+            <p>Com o intuito de compartilhar conhecimento, criei esse espaço onde você pode aproveitar e aprender sobre os mais diversos temas da tecnólogia que vão desde o Design ate a programação.</p>h1>Olá, o que vamos aprender hoje?</h1>
+            <p>Com o intuito de compartilhar conhecimento, criei esse espaço onde você pode aproveitar e aprender sobre os mais diversos temas da tecnólogia que vão desde o Design ate a programação.</p>h1>Olá, o que vamos aprender hoje?</h1>
+            <p>Com o intuito de compartilhar conhecimento, criei esse espaço onde você pode aproveitar e aprender sobre os mais diversos temas da tecnólogia que vão desde o Design ate a programação.</p>h1>Olá, o que vamos aprender hoje?</h1>
+            <p>Com o intuito de compartilhar conhecimento, criei esse espaço onde você pode aproveitar e aprender sobre os mais diversos temas da tecnólogia que vão desde o Design ate a programação.</p>h1>Olá, o que vamos aprender hoje?</h1>
+            <p>Com o intuito de compartilhar conhecimento, criei esse espaço onde você pode aproveitar e aprender sobre os mais diversos temas da tecnólogia que vão desde o Design ate a programação.</p>h1>Olá, o que vamos aprender hoje?</h1>
+            <p>Com o intuito de compartilhar conhecimento, criei esse espaço onde você pode aproveitar e aprender sobre os mais diversos temas da tecnólogia que vão desde o Design ate a programação.</p>h1>Olá, o que vamos aprender hoje?</h1>
+            <p>Com o intuito de compartilhar conhecimento, criei esse espaço onde você pode aproveitar e aprender sobre os mais diversos temas da tecnólogia que vão desde o Design ate a programação.</p>h1>Olá, o que vamos aprender hoje?</h1>
+            <p>Com o intuito de compartilhar conhecimento, criei esse espaço onde você pode aproveitar e aprender sobre os mais diversos temas da tecnólogia que vão desde o Design ate a programação.</p>h1>Olá, o que vamos aprender hoje?</h1>
+            <p>Com o intuito de compartilhar conhecimento, criei esse espaço onde você pode aproveitar e aprender sobre os mais diversos temas da tecnólogia que vão desde o Design ate a programação.</p>h1>Olá, o que vamos aprender hoje?</h1>
+            <p>Com o intuito de compartilhar conhecimento, criei esse espaço onde você pode aproveitar e aprender sobre os mais diversos temas da tecnólogia que vão desde o Design ate a programação.</p>h1>Olá, o que vamos aprender hoje?</h1>
+            <p>Com o intuito de compartilhar conhecimento, criei esse espaço onde você pode aproveitar e aprender sobre os mais diversos temas da tecnólogia que vão desde o Design ate a programação.</p>h1>Olá, o que vamos aprender hoje?</h1>
+            <p>Com o intuito de compartilhar conhecimento, criei esse espaço onde você pode aproveitar e aprender sobre os mais diversos temas da tecnólogia que vão desde o Design ate a programação.</p>h1>Olá, o que vamos aprender hoje?</h1>
+            <p>Com o intuito de compartilhar conhecimento, criei esse espaço onde você pode aproveitar e aprender sobre os mais diversos temas da tecnólogia que vão desde o Design ate a programação.</p>
+            <div class="grid-cards"></div>
+        </section>
+    </main>
+
+    <footer>
+        <div class="madeFor">
+            <span>Feito por <a href="https://denergarcia.em3de.com/" target="_blank">Dener Garcia</a>
+            </span>
+        </div>
+    </footer>
+</body>
+
+</html>
