+++ src/styles/modules/_buttons.scss
@@ -1,26 +1,81 @@
+.btn-primary {
+  min-width: 8rem;
+  width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border: none;
+  border-radius: 0.5rem;
+  padding: 0.75rem 1.5rem;
+  font-size: $fs-18;
+  font-weight: $fw-medium;
+  background-color: var(--bg-primary);
+  color: var(--bg-constrast-light);
+  box-shadow: $effect-shadow-default;
+  cursor: pointer;
 
-
-  .btn-primary {
-    min-width: 8rem;
-    width: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    border: none;
-    border-radius: 0.5rem;
-    padding: 0.75rem 1.5rem;
-    font-size: $fs-18;
-    font-weight: $fw-medium;
-    background-color: var(--bg-primary);
-    color: var(--bg-constrast-light);
-    box-shadow: $effect-shadow-default;
-    cursor: pointer;
-
-      &:hover{
-        box-shadow: $effect-shadow-hover;
-      }
-
-      &:focus{
-        box-shadow: $effect-shadow-active;
-      }
+  &:hover {
+    box-shadow: $effect-shadow-hover;
   }
+
+  &:focus {
+    box-shadow: $effect-shadow-active;
+  }
+}
+
+.btn-return {
+  display: block;
+  position: relative;
+  z-index: 10;
+  right: $fs-16;
+
+  width: $fs-40;
+  height: $fs-40;
+  border-radius: $fs-8;
+  background-color: var(--bg-primary);
+  box-shadow: $effect-shadow-default;
+  color: var(--bg-constrast-light);
+  overflow: hidden;
+  cursor: pointer;
+
+  @media screen and (max-width: 959px) {
+    position: fixed;
+  }
+
+  span:nth-child(1) {
+    position: absolute;
+    width: 20px;
+    height: 4px;
+    left: 10px;
+    top: 18px;
+    border-radius: 5rem;
+    background-color: var(--bg-constrast-light);
+    transform: translateY(0px) rotate(45deg);
+    transition: 0.2s;
+  }
+
+  span:nth-child(2) {
+    position: absolute;
+    width: 20px;
+    height: 4px;
+    left: 10px;
+    top: 18px;
+    border-radius: 5rem;
+    background-color: var(--bg-constrast-light);
+    transform: translateY(0px) rotate(-45deg);
+    transition: 0.2s;
+  }
+
+}
+
+.btn-return:hover span:nth-child(1),
+.btn-return:focus span:nth-child(1) {
+  transform: translateY(0px) rotate(45deg) scale(1.3);
+  background-color: var(--cl-feedback-error);
+}
+
+.btn-return:hover span:nth-child(2),
+.btn-return:focus span:nth-child(2) {
+  transform: translateY(0px) rotate(-45deg) scale(1.3);
+  background-color: var(--cl-feedback-error);
+}
