+++ src/styles/modules/_class-by-js.scss
@@ -11,8 +11,6 @@
     transform: translateX(0);
 }
 
-
-
 .btnMobilePressed span:nth-child(1) {
     width: 20px;
     top: 18px;
